--- backend/alembic/versions/28eefbb72165_create_mock_ticker_test_htable.py
@@ -1,62 +1,0 @@
-"""create_mock_ticker_test_HTable
-
-Revision ID: 28eefbb72165
-Revises: 80437de1f01c
-Create Date: 2026-07-04 16:06:29.766095
-"""
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel
-
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = '28eefbb72165'
-down_revision = '80437de1f01c'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('mocktickertest',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('timestamp', sa.DateTime(), nullable=False),
-    sa.Column('symbol', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-    sa.Column('price', sa.Numeric(precision=18, scale=4), nullable=False),
-    sa.Column('volume', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_mocktickertest_symbol'), 'mocktickertest', ['symbol'], unique=False)
-    op.drop_table('hist_price')
-    op.drop_index(op.f('sandbox_timescale_cache_date_idx'), table_name='sandbox_timescale_cache')
-    op.drop_table('sandbox_timescale_cache')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('sandbox_timescale_cache',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('symbol', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('price', sa.NUMERIC(precision=18, scale=4), autoincrement=False, nullable=False),
-    sa.Column('volume', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('date', postgresql.TIMESTAMP(), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('id', 'date', name=op.f('sandbox_timescale_cache_pkey'))
-    )
-    op.create_index(op.f('sandbox_timescale_cache_date_idx'), 'sandbox_timescale_cache', [sa.literal_column('date DESC')], unique=False)
-    op.create_table('hist_price',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('asset_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('volume', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('open_price', sa.NUMERIC(precision=18, scale=4), autoincrement=False, nullable=False),
-    sa.Column('high_price', sa.NUMERIC(precision=18, scale=4), autoincrement=False, nullable=False),
-    sa.Column('low_price', sa.NUMERIC(precision=18, scale=4), autoincrement=False, nullable=False),
-    sa.Column('symbol', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('offical_close', sa.NUMERIC(precision=18, scale=4), autoincrement=False, nullable=True),
-    sa.Column('date', postgresql.TIMESTAMP(), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('id', name=op.f('hist_price_pkey'))
-    )
-    op.drop_index(op.f('ix_mocktickertest_symbol'), table_name='mocktickertest')
-    op.drop_table('mocktickertest')
-    # ### end Alembic commands ###
