+++ backend/alembic/versions/635771295db8_create_mock_ticker_test_htable.py
@@ -1,0 +1,43 @@
+"""create_mock_ticker_test_HTable
+
+Revision ID: 635771295db8
+Revises: 80437de1f01c
+Create Date: 2026-07-04 16:08:43.961645
+"""
+from alembic import op
+import sqlalchemy as sa
+import sqlmodel
+
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision = '635771295db8'
+down_revision = '80437de1f01c'
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('mocktickertest',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('timestamp', sa.DateTime(), nullable=False),
+    sa.Column('symbol', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('price', sa.Numeric(precision=18, scale=4), nullable=False),
+    sa.Column('volume', sa.Integer(), nullable=False),
+    sa.PrimaryKeyConstraint('id', 'timestamp')
+    )
+    op.create_index(op.f('ix_mocktickertest_symbol'), 'mocktickertest', ['symbol'], unique=False)
+  
+    #Hyper table conversion
+    op.execute(
+        "Select create_hypertable('mocktickertest', 'timestamp', chunk_time_interval => INTERVAL '1 day');"
+    )
+
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    op.drop_index(op.f('ix_mocktickertest_symbol'), table_name='mocktickertest')
+    op.drop_table('mocktickertest')
+    # ### end Alembic commands ###
